// src/patient/patient-chart/patient-rx/models/patient-rx.model.ts
export class RxPatient {
    PatientId: string;
    Prefix: string;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    Suffix: string;
    DateOfBirth: Date;
    Gender: string;
    PhoneType: string;
    Height = 0;
    Weight = 0;
    HeightMetric = 0;
    WeightMetric = 0;
    UserId: string;
    MRN: '';
    Email: string;
    ApplicationId: string;
    Address1: string;
    Address2: string;
    City: string;
    State: string;
    PostalCode: string;
    Phone: string;
}